# router/src/main/java/fr/mightycode/cpoo/router/model/Message.java
package fr.mightycode.cpoo.router.model;

import lombok.Data;

import java.util.UUID;

@Data
public class Message {
  private UUID id;        // unique id of the message
  private long timestamp; // timestamp of the message
  private String from;    // sender address
  private String to;      // recipient address
  private String type;    // MIME type of the body
  private String body;    // body (BASE64 encoded for binary types)
}